requirements.txt
fastapi
uvicorn
openai
python-dotenv
motor
mcp
